fixes integer to decimal field

+++ main_app/models.py
@@ -3,7 +3,7 @@ from django.db import models
 class Job(models.Model):
     company = models.CharField(max_length=100)
     role = models.CharField(max_length=100)
-    salary = models.IntegerField()
+    salary = models.DecimalField(max_digits=10, decimal_places=2)
 
     def __str__(self):
         return self.company
